Restructure project to pythonic structure
